+++ SFMLConsoleApplication/SFMLConsoleApplication/Invader.h
@@ -4,7 +4,7 @@
 class Invader : public Entity 
 {
 public:	
-	Invader(float iniX, float iniY, int iniBoundryX, int iniBoundryY, float iniRad, std::string colId, float speed, sf::Texture* texture, sf::Texture* textureExplosion, Game* game);
+	Invader(int iniX, int iniY, int iniBoundryX, int iniBoundryY, float iniRad, std::string colId, float speed, sf::Texture* texture, sf::Texture* textureExplosion, Game* game);
 	~Invader();
 
 	void OnCollision(std::string CollisionId) override;
