+++ SFMLConsoleApplication/SFMLConsoleApplication/Player.h
@@ -4,7 +4,7 @@
 class Player : public Entity 
 {
 public:
-	Player(float iniX, float iniY, int boundX, int boundY, float iniRad, std::string colId, float speed, sf::Texture* texture, Game* game);
+	Player(int iniX, int iniY, int boundX, int boundY, float iniRad, std::string colId, float speed, sf::Texture* texture, Game* game);
 	~Player();
 	void OnCollision(std::string CollisionId) override;
 	void Update(float deltaTime) override;
